#283 move: 출석체크 폴더 구조 변경
